perf: move validate_and_sanitize_search_inputs
This causes all sorts of indirect imports and increases memory usage
